Merge in GITAM/dpdk from 19.11-qoriq-dev-support-non-confirmation-mode to 19.11-qoriq-dev

* commit 'd26606b9463de7cc50da1df6bd7e72859ae8c0e0':
  examples/bbdev_raw_app: Add latency test case
  examples/bbdev_raw_app: Retry enqueue operation in case of failure
  examples/bbdev_raw_app: Added confirmation mode configuration
  baseband/la12xx: Support non confirmation mode for enqueue operation
  examples/bbdev_raw_app: Check output address before updating

+++ drivers/baseband/la93xx/bbdev_la93xx.c
@@ -281,7 +281,7 @@ is_bd_ring_full(uint32_t ci, uint32_t ci_flag,
 }
 
 /* Enqueue raw operation */
-static uint16_t
+static int
 enqueue_raw_op(struct rte_bbdev_queue_data *q_data,
 	       struct rte_bbdev_raw_op *bbdev_op)
 {
